moved common fixture
